Refactoring : add Vuex in QuestionnaireCreate (#416)
When editing a questionnaire, Store the currentQuestionnaire (questionnaire being edited) in the vuex store.

This means QuestionnaireCreate and its 3 child components (QuestionnaireMetadataCreate, QuestionnaireBodyCreate, QuestionnairePreview) now share the same questionnaire object through the store. It removes the need for a lot of messaging back and forth between components, because they don't need to notify each other of changes in the questionnaire and update their copy of the questionnaire (updateXXX functions).

Overall, there is less code in the child components, it is centralized in the parent component :

move Wizard to QuestionnaireCreate, instead of duplicating it in each child component
move navigation buttons at the bottom of the page (previous, save, next) to QuestionnaireCreate, instead of duplicating them in each child component. The child components just relay the "next" and "back" messages to the parent, which handles navigation-related things.
PublishConfirmModal, which was owned by QuestionnairePreview, moved to the parent QuestionnaireCreate. This way QuestionnaireCreate owns all the saving/publishing process.
The startup is now a bit different : most of the startup in QuestionnaireCreate that was in mounted() is now moved to the watcher function of the store status. It works like this :

The store fetches the list of controls for the user and saves them in store.
QuestionnaireCreate watches for the change of store status, which signals the controls are successfully saved
QuestionnaireCreate finds the currentQuestionnaire from the list of controls, or creates a new one, and loads the whole UI.
The store now has two different load statuses : one for loading SessionUser and one for loading controls.

Co-authored-by: Sylvain Toé <sylvain@lazybird.co>