Update JS packages, and fix tests (#426)
Updated npm packages and jquery and popper which are linked with script tags.

Also fixed the tests that broke because of the update (mostly it's adding flushPromises in many tests, because the store.commit need to be finished before the expects can work).

Also added --no-source-maps to the parcel builds for questionnaire-create. That's because the new version of vuejs-datepicker breaks parcel with source maps. (maybe we could just use another datepicker... oh well.)